Dispose created views to stop the property changed notifications being sent to disposed Android views
